Moving 2d EOM to a new propagation file
